Attempt static building of core frameworks.
Uses static frameworks within Core which is then dynamically linked to the app in an attempt to improve launch times.

Scripting needs to be improved.

In order to get building you must build ReactiveCocoa with:
`$ ./Scripts/carthage-build-static.sh ReactiveCocoa --platform iOS`
and build Quick and Nimble with:
`$ carthage build Quick Nimble --platform ios`

Launch time was not increased, but that is likely due to the fact that there are so few dynamic frameworks being linked.

More investigation is required.